+++ src/main/java/codesquad/handler/UserHandler.java
@@ -15,7 +15,7 @@ import codesquad.domain.HttpMethod;
 import codesquad.domain.HttpRequest;
 import codesquad.domain.HttpResponse;
 import codesquad.domain.HttpStatus;
-import codesquad.domain.User;
+import codesquad.data.User;
 import codesquad.error.BaseException;
 import codesquad.utils.UserThreadLocal;
